Generate includes files directly from PHP
