+++ include/symbols_7_4_x64_nts.txt
@@ -1,0 +1,194 @@
+zend_long
+zend_ulong
+zend_off_t
+zend_bool
+zend_uchar
+zend_type
+ZEND_RESULT_CODE
+zend_object_handlers
+zend_class_entry
+zend_function
+zend_execute_data
+zval
+zend_refcounted
+zend_string
+zend_array
+zend_object
+zend_resource
+zend_reference
+zend_ast_ref
+zend_ast
+compare_func_t
+swap_func_t
+sort_func_t
+dtor_func_t
+copy_ctor_func_t
+zend_value
+_zval_struct
+zend_refcounted_h
+_zend_refcounted
+_zend_string
+Bucket
+HashTable
+_zend_array
+HashPosition
+HashTableIterator
+_zend_object
+_zend_resource
+zend_property_info_list
+zend_property_info_source_list
+_zend_reference
+_zend_ast_ref
+zend_ast_kind
+zend_ast_attr
+_zend_ast
+zend_ast_list
+zend_ast_zval
+zend_ast_decl
+zend_ast_process_t
+zend_intptr_t
+zend_uintptr_t
+zend_arena
+_zend_arena
+zend_op_array
+zend_op
+znode_op
+znode
+zend_ast_znode
+zend_declarables
+zend_file_context
+zend_parser_stack_elem
+user_opcode_handler_t
+_zend_op
+zend_brk_cont_element
+zend_label
+zend_try_catch_element
+zend_live_range
+zend_oparray_context
+zend_property_info
+zend_class_constant
+zend_internal_arg_info
+zend_arg_info
+zend_internal_function_info
+_zend_op_array
+zif_handler
+zend_internal_function
+_zend_function
+_zend_execute_data
+zend_closure
+std_object_handlers
+zend_object_read_property_t
+zend_object_read_dimension_t
+zend_object_write_property_t
+zend_object_write_dimension_t
+zend_object_get_property_ptr_ptr_t
+zend_object_has_property_t
+zend_object_has_dimension_t
+zend_object_unset_property_t
+zend_object_unset_dimension_t
+zend_object_get_properties_t
+zend_object_get_debug_info_t
+zend_prop_purpose
+zend_object_get_properties_for_t
+zend_object_get_method_t
+zend_object_get_constructor_t
+zend_object_dtor_obj_t
+zend_object_free_obj_t
+zend_object_clone_obj_t
+zend_object_get_class_name_t
+zend_object_compare_t
+zend_object_cast_t
+zend_object_count_elements_t
+zend_object_get_closure_t
+zend_object_get_gc_t
+zend_object_do_operation_t
+zend_object_get_t
+zend_object_set_t
+zend_object_call_method_t
+zend_object_compare_zvals_t
+_zend_object_handlers
+zend_llist_element
+llist_dtor_func_t
+llist_compare_func_t
+llist_apply_with_args_func_t
+llist_apply_with_arg_func_t
+llist_apply_func_t
+zend_llist
+zend_llist_position
+zend_encoding
+zend_stack
+zend_vm_stack
+zend_ini_entry
+zend_php_scanner_event
+_zend_vm_stack
+zend_function_entry
+zend_fcall_info
+zend_fcall_info_cache
+zend_object_iterator
+zend_object_iterator_funcs
+_zend_object_iterator
+zend_class_iterator_funcs
+_zend_serialize_data
+_zend_unserialize_data
+zend_serialize_data
+zend_unserialize_data
+zend_class_name
+zend_trait_method_reference
+zend_trait_precedence
+zend_trait_alias
+_zend_class_entry
+zend_error_handling_t
+zend_error_handling
+zend_objects_store
+_zend_ini_entry
+zend_module_entry
+zend_module_dep
+_zend_module_entry
+_zend_module_dep
+zend_gc_status
+_zend_compiler_globals
+jmp_buf
+fpu_control_t
+_zend_executor_globals
+zend_executor_globals
+executor_globals
+compiler_globals
+FILE
+zend_stream_fsizer_t
+zend_stream_reader_t
+zend_stream_closer_t
+zend_stream_type
+zend_stream
+zend_file_handle
+zend_ptr_stack
+zend_encoding_filter
+zend_lex_state
+zend_heredoc_label
+zend_ast_process
+module_registry
+zend_hash_del
+zend_hash_find
+zend_hash_add_or_update
+zend_set_user_opcode_handler
+zend_get_user_opcode_handler
+zend_do_inheritance_ex
+zend_objects_new
+zend_object_std_init
+object_properties_init
+zend_save_lexical_state
+zend_restore_lexical_state
+zend_prepare_string_for_scanning
+zend_lex_tstring
+zendparse
+zend_ast_destroy
+zend_ast_create_list_0
+zend_ast_list_add
+zend_ast_create_zval_ex
+zend_ast_create_0
+zend_ast_create_1
+zend_ast_create_2
+zend_ast_create_3
+zend_ast_create_4
+zend_ast_create_decl
+zend_register_module_ex
+zend_startup_module_ex
